% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trans.R
\name{add_trans}
\alias{add_trans}
\title{Add transparency to a color}
\usage{
add_trans(color, trans)
}
\value{

}
\description{
Add transparency to a color
}
\details{
This function adds transparancy to a color. It is from TEE. Define transparancy with an integer between 0 and 255, 0 being fully transparant and 255 being fully visible. Works with either color and trans a vector of equal length, or one of the two of length 1.
}
