% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_z.R
\name{calc_z}
\alias{calc_z}
\title{Calculate z (degree of compensation)}
\usage{
calc_z(MNPL_in, lh.params_in)
}
\arguments{
\item{MNPL_in}{User-specified value for MNPL (between 0 and 1)}

\item{lh.params_in}{a list of life history parameters}
}
\value{

}
\description{
Calculate z (degree of compensation)
}
\details{
Helper function for getting Z when user specifies MNPL
}
\examples{
(test.z <- calc_z(MNPL_in = 0.4,lh.params_in = lh.params1))
}
