% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_ce.R
\name{ce}
\alias{ce}
\title{Calculate normalized sustainable yield}
\usage{
ce(S0, S1plus, nages, K1plus, AgeMat, z, lambdaMax, E, A, P0, N0)
}
\arguments{
\item{S0}{Calf/pup survival}

\item{S1plus}{1+ survival (usually called "juvenile + adult" in literature)}

\item{nages}{Same as plus group age, called "maximum (lumped) age-class" in Annex R}

\item{K1plus}{1+ carrying capacity}

\item{AgeMat}{Age at maturity (= age at first parturition - 1)}

\item{z}{degree of compensation}

\item{E}{exploitation rate}

\item{P0}{unfished nums per recruit - 1+ adults}

\item{N0}{unfished nums per recruit - mature adults}
}
\value{
a single value of normalized yield for exploitation rate E
}
\description{
This function calculates the normalized sustainable yield, which is used to find MNPL (the population size at which productivity is maximized).
}
