% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamics.R
\name{Dynamics}
\alias{Dynamics}
\title{Generate one marine mammal population trajectory}
\usage{
Dynamics(
  S0,
  S1plus,
  K1plus,
  AgeMat,
  InitDepl,
  ConstantCatch = NA,
  ConstantF = NA,
  z,
  nyears,
  nages,
  lambdaMax
)
}
\arguments{
\item{S0}{calf survival rate}

\item{S1plus}{survival rate for animals 1 and older}

\item{K1plus}{pre-exploitation population size (1+ component of pop)}

\item{AgeMat}{age at maturity}

\item{InitDepl}{starting depletion level}

\item{ConstantCatch}{vector(length=nyears) total bycatch each year}

\item{ConstantF}{vector(length=nyears) rate of bycatch each year}

\item{z}{compensation (2.39 for bowhead whaless, set as default for others)}

\item{nyears}{number of years to simulate}

\item{nages}{"max" age, treated as plus group}

\item{lambdaMax}{max intrinsic growth rate}
}
\value{
a list containing one matrix N and one vector TotalPop
 N has dimensions nyears (rows) x nages (columns) and contains the number of individuals at each age. TotalPop, a vector of length nyears, contains all the 1+ individuals in the population.
}
\description{
Generates one trajectory for a marine mammal population, starting at InitDepl
}
