% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMYSR.R
\name{findMSYR}
\alias{findMSYR}
\title{calculate MSYR}
\usage{
findMSYR(E.start, lh.params, fmax, N0)
}
\arguments{
\item{E.start}{a starting guess for exploitation rate that will result in FMSY (num)}

\item{lh.params}{a list of life history parameters (list)}

\item{fmax}{Max theoretical fecundity (num)}

\item{unpr}{unfished numbers per recruit (num)}
}
\description{
one of the functions needed for getting MNPL. MSYR is the same as FMSY.
}
