% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_rf.R
\name{get_rf}
\alias{get_rf}
\title{Get recruitment at exploitation rate E}
\usage{
get_rf(E, S0, S1plus, nages, K1plus, AgeMat, z, A, P0, N0)
}
\arguments{
\item{E}{initial exploitation rate (num)}

\item{S0}{Calf or pup survival (num)}

\item{S1plus}{Adult survival  (num)}

\item{nages}{Number of age classes, incl plus group age (num)}

\item{K1plus}{Adult carrying capacity (num)}

\item{AgeMat}{Age at maturity (num)}

\item{z}{degree of compensation (num)}

\item{A}{Pella-Tomlinson resilience parameter (Punt 1999; Annex R). A = (FecMax - Fec0) / Fec0 (num)}

\item{P0}{Unfished 1+ numbers per recruit, \eqn{tildeP(0)}. Must double-check with AEP.}

\item{N0}{Unfished mature numbers per recruit, \eqn{tildeN(0)}. Must double-check with AEP.}
}
\value{
recruitment given exploitation rate E - to multiply by Ninit to get initial nums at age (num)
}
\description{
calculates recruitment at exploitation rate E
}
