% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFecmax2.R
\name{getFecmax2}
\alias{getFecmax2}
\title{Get max theoretical fecundity
From a ginormous bunch of algebra we did at OMF2 - IMPORTANT: when applying this calculation, use AgeMat and when defining probability of giving birth you use parturition ages (AgeMat+1)}
\usage{
getFecmax2(S0, lambdaMax, S1plus, AgeMat)
}
\arguments{
\item{S0}{Calf survival (num)}

\item{lambdaMax}{Maximum population growth rate (num)}

\item{S1plus}{Survival of age 1+ individuals (num)}

\item{AgeMat}{Age at maturity (num)}

\item{f}{'fishing' rate == bycatch rate (num)}
}
\value{
A list of numbers per recruit (npr), 1+ numbers per recruit (P1r), and numbers at age per recruit (nvec)
}
\description{
Calculate maximum theoretical fecundity Fmax
}
