% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRF.R
\name{getRF}
\alias{getRF}
\title{Get recruitment at exploitation rate E}
\usage{
getRF(E, S0, S1plus, nages, K1plus, AgeMat, z, A, unpr)
}
\arguments{
\item{E}{initial exploitation rate (num)}

\item{S0}{Calf or pup survival (num)}

\item{S1plus}{Adult survival  (num)}

\item{nages}{Number of age classes, incl plus group age (num)}

\item{K1plus}{Adult carrying capacity (num)}

\item{AgeMat}{Age at maturity (num)}

\item{z}{degree of compensation (num)}

\item{A}{Pella-Tomlinson resilience parameter (Punt 1999; Annex R). A = (FecMax - Fec0) / Fec0 (num)}

\item{unpr}{Unfished numbers-per-recruit (num)}
}
\value{
recruitment given exploitation rate E - to multiply by Ninit to get initial nums at age (num)
}
\description{
calculates recruitment at exploitation rate E
}
