% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv.logit}
\alias{inv.logit}
\title{Get inverse logit of a value}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{A number}
}
\value{
the inverse logit of x. Returns NA if is.na(x)
}
\description{
The inverse logit of x
}
\examples{
inv.logit(0)
inv.logit(1)

}
