% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_logit.R
\name{logit}
\alias{logit}
\title{Logit transform}
\usage{
logit(p)
}
\arguments{
\item{p}{Value to logit-transform}
}
\description{
Take the logit transform of a value \emph{p}
}
\examples{
x <- logit(p = 0.01)

}
