% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ptable.R
\name{make.ptable}
\alias{make.ptable}
\title{Make a performance table}
\usage{
make.ptable(traj.list, depletion, lh.params, mnpl)
}
\arguments{
\item{traj.list}{A list of trajectories from \code{Projections()}}

\item{depletion}{A vector of starting depletions (abundance relative to carrying capacity)}

\item{lh.params}{A list of life history parameters}

\item{mnpl}{Max net productivity level (defined by user)}
}
\value{

}
\description{
Make a performance table
}
