% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.proj.R
\name{plot.proj}
\alias{plot.proj}
\title{Plot projections}
\usage{
\method{plot}{proj}(
  high,
  med,
  low,
  years.plot = 50,
  ylims,
  spaghetti = FALSE,
  years.to.plot = 50,
  K1plus = 9000,
  InitDepl = 0.8,
  color.palette = c("forestgreen", "darkorange", "red"),
  multiplot = FALSE
)
}
\arguments{
\item{high}{A list containing output from Projections() (including a matrix of simulation trajectories) - corresponding to a high bycatch level (this is at the high end of the range defined by the user in the Shiny app)}

\item{med}{A list containing output from Projections() (including a matrix of simulation trajectories) - corresponding to a medium bycatch level (this is at the median of the low and high end of the range defined by the user)}

\item{low}{A list containing output from Projections() (including a matrix of simulation trajectories) - corresponding to a high bycatch level (this is at the high end of the range defined by the user)}

\item{years.plot}{- how many years to plot}

\item{spaghetti}{- either FALSE or a number, where the number is how many simulations to show from the same scenario}

\item{years.to.plot}{- how many years to plot on the x axis}

\item{multiplot}{}
}
\value{
A plot of 50 percent and 90 percent confidence intervals of population projections (if \code{spaghetti ==FALSE}) or a spaghetti plot (if \code{spaghetti ==TRUE}),  from \code{Projections()}.
}
\description{
plots outputs from several projections that result from a Projections() call.
}
