% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamics.R
\name{Dynamics}
\alias{Dynamics}
\title{Generate one marine mammal population trajectory}
\usage{
Dynamics(
  S0 = NA,
  S1plus = NA,
  K1plus = NA,
  AgeMat = NA,
  InitDepl = NA,
  ConstantCatch = NA,
  ConstantF = NA,
  z = 2.39,
  nyears = NA,
  nages = NA,
  lambdaMax = NA
)
}
\arguments{
\item{S0}{Calf survival rate}

\item{S1plus}{Survival rate for animals 1 and older}

\item{K1plus}{The pre-exploitation population size (1+ component of pop)}

\item{AgeMat}{Age at maturity}

\item{InitDepl}{Starting depletion level}

\item{ConstantCatch}{Total bycatch each year, expressed as a vector of length [nyears]}

\item{ConstantF}{vector (length=nyears) rate of bycatch each year}

\item{z}{Degree of compensation (2.39 for bowhead whales, set as default for others)}

\item{nyears}{Number of years to project}

\item{nages}{"Maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity + 2 without losing accuracy.}

\item{lambdaMax}{max intrinsic growth rate}
}
\value{
a list containing a matrix N of numbers at age (dimensions nyears (rows) x nages (columns)) and one vector TotalPop (a vector of length nyears), containing the number of age 1+ individuals in the population.
}
\description{
This function generates one trajectory for a marine mammal population, starting at a user-specified depletion level `InitDepl`.
}
\details{
The population model is a single-sex age-structured model in which the number of calves or pups born each year is density dependent, with the extent of density dependence a function of the number of mature adults \eqn{\tildeN}, the fecundity (pregnancy rate) at pre-exploitation equilibrium \deqn{f_0}, the maximum theoretical fecundity rate fmax, the degree of compensation z, and the abundance of individuals aged 1+ \deqn{ x^n + y^n = z^n } relative to carrying capacity K1+.
}
\examples{
Dynamics()

}
