% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{getF}
\alias{getF}
\title{Get the value of F that results in a given depletion level}
\usage{
getF(
  f.start,
  S0.w,
  S1plus.w,
  nages.w,
  K1plus.w,
  AgeMat.w,
  InitDepl.w,
  z.w,
  lambdaMax.w,
  unpr.w
)
}
\arguments{
\item{f.start}{initial guess for F (num)}

\item{S0.w}{Calf or pup survival. 'w' suffix indicates that z is in the wrapper fn, and is used inside the fn by optim()  (num)}

\item{S1plus.w}{Adult survival (num)}

\item{K1plus.w}{Adult carrying capacity (num)}

\item{AgeMat.w}{Age at maturity (generally AFP+1) (num)}

\item{InitDepl.w}{Depletion level to solve for (num)}

\item{z.w}{degree of compensation (num)}

\item{lambdaMax.w}{max intrinsic growth rate (num)}

\item{unpr.w}{unfished nums per recruit (num)}
}
\value{
'Fishing' (bycatch) rate that results in the depletion level provided (num)
}
\description{
solves for f that gives depletion level InitDepl.w
}
