% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_npr.R
\name{npr}
\alias{npr}
\title{Get numbers per recruit}
\usage{
npr(S0, S1plus, nages, AgeMat, f = 0)
}
\arguments{
\item{S0}{calf survival (num)}

\item{S1plus}{adult survival (num)}

\item{nages}{plus group age (num)}

\item{AgeMat}{age at maturity (num)}

\item{f}{bycatch rate (could be E; exploitation rate) (num)}
}
\value{
A list of numbers per recruit (npr), 1+ numbers per recruit (P1r), and numbers at age per recruit (nvec)
}
\description{
Calculate nums per recruit at f
}
\examples{
(unpr <- NPR(S0 = 0.9,S1plus = 0.9,AgeMat=11, nages = 13,f=0))     # unfished nums per recruit
(nprf <- NPR(S0 = 0.9,S1plus = 0.9,AgeMat=11, nages = 13,f=0.8))   # nums per recruit at F

}
