% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPR.R
\name{NPR}
\alias{NPR}
\title{Get numbers per recruit}
\usage{
NPR(S0, S1plus, nages, AgeMat, f = 0)
}
\arguments{
\item{S0}{calf survival (num)}

\item{S1plus}{adult survival (num)}

\item{nages}{plus group age (num)}

\item{AgeMat}{age at maturity (num)}

\item{f}{'fishing' rate == bycatch rate (num)}
}
\value{
A list of numbers per recruit (npr), 1+ numbers per recruit (P1r), and numbers at age per recruit (nvec)
}
\description{
Calculate nums per recruit at f
}
